+++ packages/expo-router/src/react-navigation/core/__tests__/__fixtures__/MockRouter.tsx
@@ -22,29 +22,11 @@ function getStateForRouteNamesChange(state: NavigationState, routeNames: string[
   };
 }
 
-export function MockRouter(options: DefaultRouterOptions) {
+export function MockRouter(_options: DefaultRouterOptions) {
   const router: Router<NavigationState, MockActions> = {
     type: 'test',
 
     getStateForDeclaredRoutes: BaseRouter.getStateForDeclaredRoutes,
-
-    getInitialState({ routeNames, routeParamList }) {
-      const index =
-        options.initialRouteName === undefined ? 0 : routeNames.indexOf(options.initialRouteName);
-
-      return {
-        stale: false,
-        type: 'test',
-        key: String(MockRouterKey.current++),
-        index,
-        routeNames,
-        routes: routeNames.map((name) => ({
-          name,
-          key: name,
-          params: routeParamList[name],
-        })),
-      };
-    },
 
     getRehydratedState(partialState, { routeNames, routeParamList }) {
       const state = partialState;